LayoutTests/accessibility/mac/aria-orientation-expected.txt

Implicit defaults
 
Option
Menu item 1
menubar
item one
tree item

Authored orientation
radio 1
cell
cell2
X
This test makes sure that aria-orientation works correctly and the implicit defaults are defined on different roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS slider.orientation is 'AXOrientation: AXHorizontalOrientation'
PASS combobox.orientation is 'AXOrientation: AXVerticalOrientation'
PASS listbox.orientation is 'AXOrientation: AXVerticalOrientation'
PASS menu.orientation is 'AXOrientation: AXVerticalOrientation'
PASS menubar.orientation is 'AXOrientation: AXHorizontalOrientation'
PASS tablist.orientation is 'AXOrientation: AXHorizontalOrientation'
PASS toolbar.orientation is 'AXOrientation: AXHorizontalOrientation'
PASS tree.orientation is 'AXOrientation: AXVerticalOrientation'
PASS radiogroup.orientation is 'AXOrientation: AXVerticalOrientation'
PASS treegrid.orientation is 'AXOrientation: AXHorizontalOrientation'
PASS slider2.orientation is 'AXOrientation: AXUnknownOrientation'
PASS successfullyParsed is true

TEST COMPLETE
